+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Create de-DE translation.
 - Create es-ES translation.
+- Create a HashMap with command's IDs.
 
 ### Changed
 
@@ -17,5 +18,6 @@
 
 ### Fixed
 
+- Missing variable value in JoinCommand. (Issue #15)
 - Wrong translation key in LoopComponent. (Issue #16)
 - Wrong translations variables in HydrogenManager. (Issue #16)